Extract Excel cell helpers
